// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Priority = {
  low: 'low',
  medium: 'medium',
  high: 'high'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const GoalStatus = {
  active: 'active',
  completed: 'completed',
  abandoned: 'abandoned'
} as const

export type GoalStatus = (typeof GoalStatus)[keyof typeof GoalStatus]


export const Frequency = {
  daily: 'daily',
  weekly: 'weekly'
} as const

export type Frequency = (typeof Frequency)[keyof typeof Frequency]